Add receive, buy and sell deep links
Support /receive, /buy and /sell under gem:// and gemwallet.com, with an
optional asset for receive and an optional whole-dollar fiat amount for
buy and sell.

Register the gem scheme for every host on Android instead of gem://wc
only, so deep links other than WalletConnect reach the app, and route
support chat links through the app before falling back to a custom tab.

Build the presented asset input from the deep linked asset instead of the
chain native asset, so a token link opens that token.

+++ android/app/src/main/kotlin/com/gemwallet/android/ui/navigation/RouteArgumentsNavEntryDecorator.kt
@@ -46,7 +46,7 @@ internal fun routeArguments(vararg arguments: Pair<RouteArgument, Any?>): Map<St
 internal fun assetIdArgument(assetId: AssetId): Pair<RouteArgument, String> =
     RouteArgument.AssetId to assetId.toIdentifier()
 
-internal fun fiatAmountArgument(amount: Double?): Pair<RouteArgument, Double?> =
+internal fun fiatAmountArgument(amount: Int?): Pair<RouteArgument, Int?> =
     RouteArgument.FiatAmount to amount
 
 internal fun contactIdArgument(contactId: String): Pair<RouteArgument, String> =
